trying to fix up attributes

+++ NES/Core/PPUNES.h
@@ -108,7 +108,7 @@ private:
     // Background pattern shift registers
     uint16_t m_bgPatternShift0;
     uint16_t m_bgPatternShift1;
-    uint16_t m_bgPalletteShift0;
+    uint8_t m_bgPalletteShift0;
     uint8_t m_bgPalletteShift1;
     
     // Emulation
